- unnecessary type casting removed

+++ src/draw/polygons/polygon.rs
@@ -29,7 +29,7 @@ pub fn render(ctx: &mut GraphContext, props: &PolygonProperties) {
     }
 
     let angle_step = 2.0 * PI / props.num_sides as f64; // Angle between each vertex
-    let rotation_radians = props.rotation_angle as f64 * PI / 180.0; // Convert rotation angle to radians
+    let rotation_radians = props.rotation_angle * PI / 180.0; // Convert rotation angle to radians
 
     // Calculating all vertex positions
     let mut vertices = Vec::new();
